+++ Tests/Integration/Help.tests.ps1
@@ -46,7 +46,6 @@ BeforeDiscovery {
         'PSFileTransfer',
         'AutomatedLabDefinition',
         'AutomatedLabWorker',
-        'HostsFile',
         'AutomatedLabNotifications',
         'AutomatedLabCore'
     )